+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Compatibility with TYPO3 v9 LTS (#41)
 - Signal/slots in favour of PSR-14 events (#43)
+- Deprecated methods AbstractType->isEmpty() and SchemaManager->setMainEntityOfWebPage() (#44)
+- Deprecated class TypesProvider (#44)
 
 ## [1.11.1] - 2021-04-06